Conflicts:
	src/main/resources/application.properties

+++ src/main/java/com/mycompany/dao/PlacementDao.java
@@ -22,7 +22,7 @@ public interface PlacementDao {
     List<Placement> findByIdLandlord(int idLandlord);
     List<Placement> findByComforts(Set<Comforts> comfortses);
     List<Placement> findByParametr(Search search);
-    Integer findNumberByParametr(Search search);
+    Long findNumberByParametr(Search search);
     Placement findById(int id);
     Placement save(Placement placement);
     boolean delete(Placement placement);
